+++ packages/bundle-source/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [3.5.1](https://github.com/endojs/endo/compare/@endo/bundle-source@3.5.0...@endo/bundle-source@3.5.1) (2025-01-24)
+
+**Note:** Version bump only for package @endo/bundle-source
+
+
+
+
+
 ## [3.5.0](https://github.com/endojs/endo/compare/@endo/bundle-source@3.4.2...@endo/bundle-source@3.5.0) (2024-11-13)