+++ packages/promise-kit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [1.1.9](https://github.com/endojs/endo/compare/@endo/promise-kit@1.1.8...@endo/promise-kit@1.1.9) (2025-01-24)
+
+**Note:** Version bump only for package @endo/promise-kit
+
+
+
+
+
 ### [1.1.8](https://github.com/endojs/endo/compare/@endo/promise-kit@1.1.7...@endo/promise-kit@1.1.8) (2024-11-13)
 
 **Note:** Version bump only for package @endo/promise-kit