+++ packages/test262-runner/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.1.44](https://github.com/endojs/endo/compare/@endo/test262-runner@0.1.43...@endo/test262-runner@0.1.44) (2025-01-24)
+
+**Note:** Version bump only for package @endo/test262-runner
+
+
+
+
+
 ### [0.1.43](https://github.com/endojs/endo/compare/@endo/test262-runner@0.1.42...@endo/test262-runner@0.1.43) (2024-11-13)
 
 **Note:** Version bump only for package @endo/test262-runner